backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.15.1
MarkupSafe==2.1.5
setuptools==69.2.0
sqlparse==0.5.0
Werkzeug==3.0.2
wheel==0.43.0
